// lib/domain/main.types.ts
export type YoutubeNotificationProcessingMode = 'IMMEDIATE' | 'SCHEDULED' | 'SKIP'
export type YoutubeCaptionType = 'AUTO_GENERATED' | 'USER_GENERATED' | 'NONE'

export interface TimeStamps {
    createdAt: number
    updatedAt: number
}

export type AcceptablePKValue =
    | `${AcceptablePK.YOUTUBE_VIDEO}/${string}`
    | `${AcceptablePK.YOUTUBE_NOTIFICATION}/${string}`
    | `${AcceptablePK.TRANSCRIPTION}/${string}`
    | `${AcceptablePK.SUMMARY}/${string}`
    | AcceptablePK.SUBSCRIBED_CHANNEL
    | AcceptablePK.TELEGRAM_CHANNEL

export interface MainItem extends TimeStamps {
    [MainTable.PK]: AcceptablePKValue
    [MainTable.SK]: string
    expireAt?: number
}
// ...
export interface YoutubeVideoItem extends MainItem {
    [MainTable.PK]: `${AcceptablePK.YOUTUBE_VIDEO}/${string}`
    [MainTable.SK]: string
    [MainTable.TIMESTAMP]: number
    [VIDEO_TYPE_KEY]: YoutubeVideoType
    [VIDEO_GENRE_KEY]: Exclude<VideoGenre, 'ALARM'>
    videoId: string
    channelId: string
    videoTitle: string
    channelTitle: string
    channelUri: string
    publishedAt: number
    captions: YoutubeCaptionType
    processingMode: YoutubeNotificationProcessingMode
}
// ...
export type YoutubeVideoType = 'SHORT' | 'STANDARD' | 'LIVE' | 'UNKNOWN' | 'UPCOMING' | 'LONG'

export type VideoGenre = 'TINFOIL' | 'SOFTWARE_ENGINEERING' | 'ALARM' | 'POLITICS' | 'SCIENCE'

export type ErrorOutput = Readonly<{
    [ERROR_OUTPUT_ATTR_KEY]: string
    payload: object
    subject: string
}>

export type ChatNewsMessageInputPayload = Readonly<{
    [VIDEO_GENRE_KEY]: VideoGenre
    message: string
}>

export type Message = Readonly<{
    type: MessageType
    createdAt: number
    sendAt: number
}>

export type YoutubeNotification = Readonly<{
    videoId: string
    channelId: string
    videoTitle: string
    channelTitle: string
    channelUri: string
    publishedAt: number
    processingMode: YoutubeNotificationProcessingMode
    captions: YoutubeCaptionType
    genre: Exclude<VideoGenre, 'ALARM'>
    [VIDEO_TYPE_KEY]: YoutubeVideoType
}>